// frontend/src/types/settings.types.ts
export interface userSettingType {
  username: string;
  name: string;
  role: string;
  password: string;
}

export interface taxListType {
  name: string;
  value: string;
  _id: string;
}

export interface CompanyDataType {
  name: string;
  address: string;
  website: string;
  phoneOne: string;
  phoneTwo: string;
}